feat(editorconfig) add text editor settings
- Add settings for text editors with EditorConfig.

“EditorConfig helps developers define and maintain consistent coding
styles between different editors and IDEs. The EditorConfig project
consists of a file format for defining coding styles and a collection
of text editor plugins that enable editors to read the file format and
adhere to defined styles. EditorConfig files are easily readable and
they work nicely with version control systems.”

More details:
- http://editorconfig.org
- https://github.com/editorconfig